Release of version 1.0 install.py is the installer which installs ignis.exe
