+++ README.md
@@ -1,1 +1,25 @@
-# Bright
+# Ignis
+Ignis is a lightweight application designed to give you precise control over your screen brightness. Unlike the default Windows functionality, which adjusts brightness in increments of 10%, this tool allows you to increase or decrease brightness by just 1%, offering greater precision and a more customizable user experience.
+
+## Installation:
+
+1. Clone the repository:
+```
+git clone :https://github.com/BlueFlareJinx/Ignis.git
+```
+
+2. Navigate to the project directory:
+```
+cd Ignis
+```
+
+3. Run the installer to install Ignis:
+```
+py install.py
+```
+
+**After this, the application will automatically start when you restart your laptop.**
+
+## Usage:
+* Increase Brightness by 1%: ```Alt + Shift + a```
+* Decrease Brightness by 1%: ```Alt + Shift + z```
